requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-wtf==0.15.1
email-validator==1.1.3
werkzeug==2.0.1
flask-sqlalchemy==2.5.1
python-dotenv==0.19.0 